backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
qdrant-client
sentence-transformers
PyMuPDF
numpy
google-generativeai
scikit-learn
python-dotenv
pdf2image
pytesseract
httpx
trafilatura
